Add files via upload
This is an assignment given by Bluejay Delivery for a Software Development internship. In this assignment, I was asked to read an Excel file(.xlsx) and take the file as an input.  As a second step, I had to Programmatically analyze the file and print in the console the name and position of employees 
      a) who have worked for 7 consecutive days.
      b) who have less than 10 hours of time between shifts but greater than 1 hour
      c) Who has worked for more than 14 hours in a single shift.

I successfully completed the project and gave my cent percent to meet the requirements of the given task.